requirements.txt
yt-dlp==2023.12.30
flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
Pillow==10.1.0
python-dotenv==1.0.0
